Relocate testing functions that are only used in one place
